+++ relay-config/src/config.rs
@@ -912,6 +912,16 @@ fn spool_disk_usage_refresh_frequency_ms() -> u64 {
     100
 }
 
+/// Default bounded buffer size for handling backpressure.
+fn spool_max_backpressure_envelopes() -> usize {
+    500
+}
+
+/// Default max memory usage for unspooling.
+fn spool_max_backpressure_memory_percent() -> f32 {
+    0.9
+}
+
 /// Persistent buffering configuration for incoming envelopes.
 #[derive(Debug, Serialize, Deserialize)]
 pub struct EnvelopeSpool {
@@ -955,6 +965,20 @@ pub struct EnvelopeSpool {
     /// internal page stats.
     #[serde(default = "spool_disk_usage_refresh_frequency_ms")]
     disk_usage_refresh_frequency_ms: u64,
+    /// The amount of envelopes that the envelope buffer can push to its output queue.
+    #[serde(default = "spool_max_backpressure_envelopes")]
+    max_backpressure_envelopes: usize,
+    /// The relative memory usage above which the buffer service will stop dequeueing envelopes.
+    ///
+    /// Only applies when [`Self::path`] is set.
+    /// This value should be lower than [`Health::max_memory_percent`] to prevent flip-flopping.
+    ///
+    /// Warning: this threshold can cause the buffer service to deadlock when the buffer itself
+    /// is using too much memory (influenced by [`Self::max_batches`] and [`Self::disk_batch_size`]).
+    ///
+    /// Defaults to 90% (5% less than max memory).
+    #[serde(default = "spool_max_backpressure_memory_percent")]
+    max_backpressure_memory_percent: f32,
     /// Version of the spooler.
     #[serde(default)]
     version: EnvelopeSpoolVersion,
@@ -991,6 +1015,8 @@ impl Default for EnvelopeSpool {
             max_batches: spool_envelopes_stack_max_batches(),
             max_envelope_delay_secs: spool_envelopes_max_envelope_delay_secs(),
             disk_usage_refresh_frequency_ms: spool_disk_usage_refresh_frequency_ms(),
+            max_backpressure_envelopes: spool_max_backpressure_envelopes(),
+            max_backpressure_memory_percent: spool_max_backpressure_memory_percent(),
             version: EnvelopeSpoolVersion::default(),
         }
     }
@@ -2210,6 +2236,16 @@ impl Config {
     /// Returns the refresh frequency for disk usage monitoring as a [`Duration`] object.
     pub fn spool_disk_usage_refresh_frequency_ms(&self) -> Duration {
         Duration::from_millis(self.values.spool.envelopes.disk_usage_refresh_frequency_ms)
+    }
+
+    /// Returns the maximum number of envelopes that can be put in the bounded buffer.
+    pub fn spool_max_backpressure_envelopes(&self) -> usize {
+        self.values.spool.envelopes.max_backpressure_envelopes
+    }
+
+    /// Returns the relative memory usage up to which the disk buffer will unspool envelopes.
+    pub fn spool_max_backpressure_memory_percent(&self) -> f32 {
+        self.values.spool.envelopes.max_backpressure_memory_percent
     }
 
     /// Returns the maximum size of an event payload in bytes.